connector choice in iltm
Former-commit-id: 26d3665897315ad4c905306e4d6f31510e058ecf [formerly 5876a51ae5c595459403d31474a9b117119651f3]
Former-commit-id: cfa4730e0fb78b061d4dc844c493a437777dc952

+++ dtapy/assignment.py
@@ -89,7 +89,7 @@ class Assignment:
         turns = self.__build_turns(tot_nodes, nodes)
         print("turns passed")
 
-        link_capacity = np.array([d['capacity'] for (_, _, d) in sorted_edges], dtype=np.uint32)
+        link_capacity = np.array([d['capacity'] for (_, _, d) in sorted_edges], dtype=np.float32)
         free_speed = np.array([d['free_speed'] for (_, _, d) in sorted_edges], dtype=np.float32)
         lanes = np.array([d['lanes'] for (_, _, d) in sorted_edges], dtype=np.uint8)
         length = np.array([d['length'] for (_, _, d) in sorted_edges], dtype=np.float32)
